perform accurate initial wave position

+++ battle/attack/AttackSimple.java
@@ -122,7 +122,7 @@ public class AttackSimple extends AttackAb {
 		if (capt.size() > 0 && proc.WAVE.exists()) {
 			int dire = model.getDire();
 			int wid = dire == 1 ? W_E_WID : W_U_WID;
-			int addp = (dire == 1 ? W_E_INI : W_U_INI) + wid / 2;
+			double addp = (dire == 1 ? W_E_INI : W_U_INI) + wid / 2.0;
 			double p0 = model.getPos() + dire * addp;
 			// generate a wave when hits somebody
 
@@ -136,7 +136,7 @@ public class AttackSimple extends AttackAb {
 		if(capt.size() > 0 && proc.MINIWAVE.exists()) {
 			int dire = model.getDire();
 			int wid = dire == 1 ? W_E_WID : W_U_WID;
-			int addp = (dire == 1 ? W_E_INI : W_U_INI) + wid / 2;
+			double addp = (dire == 1 ? W_E_INI : W_U_INI) + wid / 2.0;
 			double p0 = model.getPos() + dire * addp;
 
 			ContWaveDef wave = new ContWaveDef(new AttackWave(attacker, this, p0, wid, WT_MINI), p0, layer, false);
